Fix test builds (#109)
* Remove old serializers (deprecated after v3)
* Add sanity tests to NETFramework project
* Fix package hint path
* Remove old dependencies
* Reuse test code
* Moving tests to Common. Adding them to test projects as links
